+++ package.json
@@ -29,6 +29,7 @@
     "react-hook-form": "^7.73.1",
     "react-lottie-player": "^2.1.0",
     "tailwind-merge": "^3.5.0",
+    "zod": "^4.4.3",
     "zustand": "^5.0.12"
   },
   "devDependencies": {
