+++ src/app/(private)/mypage/reservation-status/page.tsx
@@ -2,8 +2,14 @@
 
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
-import ReservationCalendar from '@/app/(private)/mypage/reservation-status/components/ReservationCalendar';
-import ReservationPanel from '@/app/(private)/mypage/reservation-status/components/ReservationPanel';
+import {
+  MOCK_ACTIVITIES,
+  MOCK_CALENDAR_SCHEDULES,
+  MOCK_PANEL_SCHEDULES,
+  MOCK_RESERVATIONS,
+} from '@/features/my-page/reservation-status/mocks/index';
+import ReservationCalendar from '@/features/my-page/reservation-status/ui/ReservationCalendar';
+import ReservationPanel from '@/features/my-page/reservation-status/ui/ReservationPanel';
 import SelectDropdown from '@/shared/ui/dropdown/select/SelectDropdown';
 import SelectDropdownContent from '@/shared/ui/dropdown/select/SelectDropdownContent';
 import SelectDropdownItem from '@/shared/ui/dropdown/select/SelectDropdownItem';
@@ -11,12 +17,6 @@ import SelectDropdownTrigger from '@/shared/ui/dropdown/select/SelectDropdownTri
 import SelectDropdownValue from '@/shared/ui/dropdown/select/SelectDropdownValue';
 import EmptyState from '@/shared/ui/empty-state/EmptyState';
 import PageHeader from '@/shared/ui/page-header/PageHeader';
-import {
-  MOCK_ACTIVITIES,
-  MOCK_CALENDAR_SCHEDULES,
-  MOCK_PANEL_SCHEDULES,
-  MOCK_RESERVATIONS,
-} from './mocks';
 
 const toDateStr = (date: Date) =>
   `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`;
