+++ src/features/my-page/reservation-status/ui/ReservationCard.tsx
@@ -1,4 +1,4 @@
-import type { Reservation } from '@/app/(private)/mypage/reservation-status/types/reservation';
+import type { Reservation } from '@/features/my-page/reservation-status/types/reservation';
 import Button from '@/shared/ui/button/Button';
 import StateBadge from '@/shared/ui/state-badge/StateBadge';
