+++ src/features/my-page/reservation-status/ui/ReservationPanelContent.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { useState } from 'react';
-import ReservationCard from '@/app/(private)/mypage/reservation-status/components/ReservationCard';
 import type {
   Schedule,
   Reservation,
   ReservationStatus,
-} from '@/app/(private)/mypage/reservation-status/types/reservation';
+} from '@/features/my-page/reservation-status/types/reservation';
+import ReservationCard from '@/features/my-page/reservation-status/ui/ReservationCard';
 import { DeleteIcon } from '@/shared/assets/icons';
 import SelectDropdown from '@/shared/ui/dropdown/select/SelectDropdown';
 import SelectDropdownContent from '@/shared/ui/dropdown/select/SelectDropdownContent';
